--- app/src/main/res/layout/fragment_safe_overview.xml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/safe_overview_root"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:background="@color/background" >
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/toolbar"
-        style="@style/Toolbar"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" >
-
-        <pm.gnosis.blockies.BlockiesImageView
-            android:id="@+id/safe_image"
-            android:layout_width="@dimen/safe_blockie_size"
-            android:layout_height="@dimen/safe_blockie_size"
-            android:layout_marginStart="16dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/safe_name"
-            style="@style/ToolbarTitle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:visibility="visible"
-            app:layout_constraintBottom_toTopOf="@+id/safe_address"
-            app:layout_constraintEnd_toStartOf="@+id/safe_selection"
-            app:layout_constraintStart_toEndOf="@+id/safe_image"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="My Company Funds" />
-
-        <TextView
-            android:id="@+id/safe_address"
-            style="@style/ToolbarSubtitle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@+id/safe_selection"
-            app:layout_constraintStart_toEndOf="@+id/safe_image"
-            app:layout_constraintTop_toBottomOf="@+id/safe_name"
-            tools:text="0xAB...146F" />
-
-        <ImageView
-            android:id="@+id/safe_selection"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="16dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@drawable/ic_circle_dropdown_24dp" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <View
-        android:id="@+id/toolbar_shadow"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/toolbar_shadow_size"
-        android:background="@drawable/toolbar_dropshadow"
-        app:layout_constraintTop_toBottomOf="@id/toolbar" />
-
-    <fragment
-        android:id="@+id/safe_overview_content"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:defaultNavHost="true"
-        app:layout_constraintBottom_toTopOf="@id/bottom_navigation"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbar_shadow"
-        app:navGraph="@navigation/safe_overview_nav" />
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottom_navigation"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        app:itemBackground="@color/white"
-        app:itemIconTint="@drawable/bottom_nav_item"
-        app:itemTextColor="@drawable/bottom_nav_item"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:menu="@menu/bottom_nav_menu" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
