+++ app/src/main/res/layout/toolbar_safe_overview.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/Toolbar">
+
+    <pm.gnosis.blockies.BlockiesImageView
+        android:id="@+id/safe_image"
+        android:layout_width="@dimen/safe_blockie_size"
+        android:layout_height="@dimen/safe_blockie_size"
+        android:layout_marginStart="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/safe_name"
+        style="@style/ToolbarTitle"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:visibility="visible"
+        app:layout_constraintBottom_toTopOf="@+id/safe_address"
+        app:layout_constraintEnd_toStartOf="@+id/safe_selection"
+        app:layout_constraintStart_toEndOf="@+id/safe_image"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="My Company Funds" />
+
+    <TextView
+        android:id="@+id/safe_address"
+        style="@style/ToolbarSubtitle"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toStartOf="@+id/safe_selection"
+        app:layout_constraintStart_toEndOf="@+id/safe_image"
+        app:layout_constraintTop_toBottomOf="@+id/safe_name"
+        tools:text="0xAB...146F" />
+
+    <ImageView
+        android:id="@+id/safe_selection"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:srcCompat="@drawable/ic_circle_dropdown_24dp" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
+
+
+
+
+
