--- app/src/main/res/navigation/safe_overview_nav.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/safe_overview_nav_graph"
-    app:startDestination="@id/noSafeFragment">
-
-    <fragment
-        android:id="@+id/noSafeFragment"
-        android:name="io.gnosis.safe.ui.safe.NoSafeFragment"
-        android:label="NoSafeFragment"
-        tools:layout="@layout/fragment_no_safes" >
-
-        <action
-            android:id="@+id/action_noSafeFragment_to_safeBalancesFragment"
-            app:destination="@id/safeBalancesFragment" />
-
-    </fragment>
-
-    <fragment
-        android:id="@+id/safeBalancesFragment"
-        android:name="io.gnosis.safe.ui.safe.SafeBalancesFragment"
-        android:label="SafeBalancesFragment"
-        tools:layout="@layout/fragment_safe_balances" />
-
-    <fragment
-        android:id="@+id/safeSettingsFragment"
-        android:name="io.gnosis.safe.ui.safe.settings.SafeSettingsFragment"
-        android:label="SafeSettingsFragment" >
-
-        <action
-            android:id="@+id/action_safeSettingsFragment_to_noSafeFragment"
-            app:destination="@id/noSafeFragment" />
-
-    </fragment>
-
-</navigation>
