+++ app/src/main/res/values/dimens.xml
@@ -3,4 +3,5 @@
     <dimen name="toolbar_shadow_size">1dp</dimen>
     <dimen name="primary_button_radius">16dp</dimen>
     <dimen name="safe_blockie_size">36dp</dimen>
+    <dimen name="nav_bar_height">56dp</dimen>
 </resources>
