+++ buildsystem/versions.gradle
@@ -10,7 +10,7 @@ ext {
             // Runtime dependencies
             androidx_app_compat       : '1.1.0',
             androidx_card_view        : '1.0.0',
-            androidx_constraint_layout: '1.1.3',
+            androidx_constraint_layout: '2.0.0-beta4',
             androidx_lifecycle        : '2.2.0',
             androidx_test_ext         : '1.1.0',
             androidx_recycler_view    : '1.1.0',
